Move errors to new package
