+++ tts/src/browserMain/kotlin/nl/marc_apps/tts/TextToSpeech.kt
@@ -1,6 +1,7 @@
 package nl.marc_apps.tts
 
 import kotlinx.browser.window
+import nl.marc_apps.tts.errors.TextToSpeechNotSupportedError
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
